Fix a bug by closing the dialog after editing elements
